Camera - ConstBuffer에 업데이트 시, winSize를 uint가 아닌 float로 업데이트 하도록 함 #37

+++ SOCEngine/SOCEngine/Rendering/Camera/CameraConstBuffer.h
@@ -10,7 +10,7 @@ namespace Rendering
 		{
 			Math::Vector4 viewPos;
 			float clippingNear, clippingFar;
-			Math::Size<unsigned int> screenSize;
+			Math::Size<float> screenSize;
 		};
 	}
 }
